+++ driver/web/rest/apis.go
@@ -607,6 +607,36 @@ func (h ApisHandler) GetExtBuildingAccess(w http.ResponseWriter, r *http.Request
 	w.Write(data)
 }
 
+//GetUserByIdentifier gets an user by identifier
+// @Description Gets an user by identifier
+// @Tags External
+// @ID GetUserByIdentifier
+// @Param identifier query string true "Identifier"
+// @Success 200
+// @Security ExternalAuth
+// @Router /covid19/external/user [get]
+func (h ApisHandler) GetUserByIdentifier(w http.ResponseWriter, r *http.Request) {
+	externalIDKeys, ok := r.URL.Query()["identifier"]
+	if !ok || len(externalIDKeys[0]) < 1 {
+		log.Println("external key is missing")
+		return
+	}
+	identifier := externalIDKeys[0]
+
+	user, err := h.app.Services.GetUser(identifier)
+	if err != nil {
+		log.Printf("Error on getting user by identifier and last name %s", err)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	if user == nil {
+		http.Error(w, "There is no user with that identiier", http.StatusNotFound)
+		return
+	}
+	w.WriteHeader(http.StatusOK)
+}
+
 //GetCounty gets a county
 // @Description Gets a county
 // @Tags Covid19
@@ -1671,6 +1701,7 @@ func (h ApisHandler) GetExtJoinExternalApproval(current model.User, account mode
 	result := make([]joinGroupExtApprovement, len(items))
 	for i, c := range items {
 		result[i] = joinGroupExtApprovement{ID: c.ID, GroupName: c.GroupName, FirstName: c.FirstName, LastName: c.LastName,
+			Email: c.Email, Phone: c.Phone,
 			DateCreated: c.DateCreated, ExternalApproverID: c.ExternalApproverID, ExternalApproverLastName: c.ExternalApproverLastName,
 			Status: c.Status}
 	}
@@ -2729,6 +2760,35 @@ func (h ApisHandler) GetRosterByPhone(appVersion *string, w http.ResponseWriter,
 		return
 	}
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(data)
+}
+
+//GetTime gives the current time in UTC
+// @Description Gives the current time in UTC.
+// @Tags Covid19
+// @ID GetTime
+// @Accept json
+// @Success 200 {object} timeResponse
+// @Security RokwireAuth
+// @Router /covid19/time [get]
+func (h ApisHandler) GetTime(appVersion *string, w http.ResponseWriter, r *http.Request) {
+	time, err := h.app.Services.GetTime()
+	if err != nil {
+		log.Printf("error on getting current time - %s", err)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	response := timeResponse{Time: *time}
+	data, err := json.Marshal(response)
+	if err != nil {
+		log.Println("Error on marshal current time")
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json; charser=utf-8")
 	w.WriteHeader(http.StatusOK)
 	w.Write(data)
 }