+++ driver/web/rest/main.go
@@ -202,11 +202,17 @@ type joinGroupExtApprovement struct {
 	GroupName                string    `json:"group_name"`
 	FirstName                string    `json:"first_name"`
 	LastName                 string    `json:"last_name"`
+	Email                    *string   `json:"email"`
+	Phone                    *string   `json:"phone"`
 	DateCreated              time.Time `json:"date_created"`
 	ExternalApproverID       string    `json:"external_approver_id"`
 	ExternalApproverLastName string    `json:"external_approver_last_name"`
 	Status                   string    `json:"status"`
 } // @name JoinGroupExtApprovement
+
+type timeResponse struct {
+	Time time.Time `json:"time"`
+} // @name Time
 
 func convertToDaysOfOperations(list []locationOperationDayRequest) []model.OperationDay {
 	var doo []model.OperationDay